move switzerland to switzerland.ch
