+++ switzerland/src/main/java/org/eqasim/switzerland/ch/ovgk/OVGK.java
@@ -1,0 +1,5 @@
+package org.eqasim.switzerland.ch.ovgk;
+
+public enum OVGK {
+	A, B, C, D, None
+}
