+++ switzerland/src/main/java/org/eqasim/switzerland/ch/ovgk/OVGKConstants.java
@@ -1,0 +1,11 @@
+package org.eqasim.switzerland.ch.ovgk;
+
+public final class OVGKConstants {
+	private OVGKConstants() {
+	}
+
+	static public final String STOP_CATEGORY_ATTRIBUTE = "stopCategory";
+
+	static public final int WORST_STOP_CATEGORY = 5;
+	static public final OVGK BEST_OVGK = OVGK.A;
+}
