--- switzerland/src/main/java/org/eqasim/switzerland/ovgk/OVGK.java
@@ -1,5 +1,0 @@
-package org.eqasim.switzerland.ovgk;
-
-public enum OVGK {
-	A, B, C, D, None
-}
