--- switzerland/src/main/java/org/eqasim/switzerland/ovgk/OVGKConstants.java
@@ -1,11 +1,0 @@
-package org.eqasim.switzerland.ovgk;
-
-public final class OVGKConstants {
-	private OVGKConstants() {
-	}
-
-	static public final String STOP_CATEGORY_ATTRIBUTE = "stopCategory";
-
-	static public final int WORST_STOP_CATEGORY = 5;
-	static public final OVGK BEST_OVGK = OVGK.A;
-}
